refactor: reorganize codebase to reflect five-stage pipeline architecture
Restructure the codebase to clearly align with the five stages described
in the README, improving code organization and making the architecture
more intuitive for developers.

**Module Renames:**
- pipeline.py → discovery.py (Stage 1: Discovery)
- checks.py → detection.py (Stage 3: Detection)
- lineage.py → linking.py (Stage 4: Linking)
- static_generator.py → generation.py (Stage 5: Generation)
- extractor.py remains unchanged (Stage 2: Extraction)

**Updated Components:**
- All import statements across codebase
- README.md with consistent five-stage terminology
- Test files to use new module names
- CLI commands documentation
- Project structure diagrams

**Documentation Improvements:**
- Clear mapping of each module to its pipeline stage
- Consistent use of five-stage terminology throughout
- Updated data flow diagram
- Enhanced extension guide with stage references

**Benefits:**
- Clearer code organization matching documentation
- Easier to locate code for specific pipeline stages
- Better onboarding experience for new contributors
- Improved maintainability and extensibility

All imports verified and CLI functionality tested.
