+++ src/mandate_pipeline/__init__.py
@@ -2,8 +2,8 @@
 
 from .downloader import download_document
 from .extractor import extract_text, extract_operative_paragraphs, extract_lettered_paragraphs
-from .checks import load_checks, run_checks
-from .pipeline import load_patterns, generate_symbols, discover_documents
+from .detection import load_checks, run_checks
+from .discovery import load_patterns, generate_symbols, discover_documents
 
 __all__ = [
     "download_document",
