+++ src/mandate_pipeline/detection.py
