+++ src/mandate_pipeline/discovery.py
